requirements.txt
pillow
python-decouple
numpy==1.9.2
celery
redis
pandas
# pypiwin32
pytesseract



